LndClient.Data.LndEnv.readEnv for simplicity of library usage in some cases
